Splitting Logik into Models, Controllers, Routes
